GRHB-430: + user mentoring permissions when becoming a mentor

+++ backend/src/services/group/group.service.ts
@@ -1,4 +1,8 @@
-import { ExceptionMessage, StringCase } from '~/common/enums/enums';
+import {
+  ExceptionMessage,
+  ProtectedGroupKey,
+  StringCase,
+} from '~/common/enums/enums';
 import {
   EntityPagination,
   EntityPaginationRequestQueryDto,
@@ -152,6 +156,18 @@ class Group {
     const { id, groupsRequestDto } = data;
     const { name, permissionIds, userIds } = groupsRequestDto;
 
+    const groupById = await this.#groupsRepository.getById(id);
+
+    const isProtectedGroup =
+      groupById &&
+      Object.values(ProtectedGroupKey).some((key) => key === groupById.key);
+
+    if (isProtectedGroup) {
+      throw new GroupsError({
+        message: ExceptionMessage.PROTECTED_GROUP_UPDATE,
+      });
+    }
+
     const group = await this.#groupsRepository
       .update({
         id,
@@ -191,9 +207,25 @@ class Group {
   }
 
   public async delete(id: number): Promise<boolean> {
+    const groupById = await this.#groupsRepository.getById(id);
+
+    const isProtectedGroup =
+      groupById &&
+      Object.values(ProtectedGroupKey).some((key) => key === groupById.key);
+
+    if (isProtectedGroup) {
+      throw new GroupsError({
+        message: ExceptionMessage.PROTECTED_GROUP_UPDATE,
+      });
+    }
+
     const deletedGroupsCount = await this.#groupsRepository.delete(id);
 
     return Boolean(deletedGroupsCount);
+  }
+
+  public getByKey(key: string): Promise<GroupsItemResponseDto | null> {
+    return this.#groupsRepository.getByKey(key);
   }
 }
